Add files via upload
Version 2.1.6 Uses Mac OS X system menu and GroupBox class which is a JInternalFrame without a title bar.  VoiceOver sees it as a "Group" and gives the proper instructions.  Everything about the VfoDisplayControl is lined up without using layout managers.
The math is all there to calculate the sizes if the main JFrame is expanded.  It is set as non adjustable now.  Lance, let me know if this runs ok on your box.  I ran it from command just fine on my box.