Add files via upload
Fixed that nasty problem where voicedOver would not read the JTabbedPane tabs as they were selected.  The fix was to update the project to JDK14 after spending many days removing things to find the problem.  This is version 5.0.10 and there are known bugs, especially in the JRX scan and scope functions.