Add files via upload
Have added the first substitute for a ComboBox.  VoiceOver does not work well with the JComboBox on the MacOS.  So will replace all the JComboBox instances with the 
ListButton class which opens a List in a Dialog.  The JRX_TX class is RWListButton.  It works well with voiceOver and the instructions that voiceOver speaks actually work.  The first implementation is the CtcssListButton which extends RWListButton.  Since the CTCSS tones are well established, it initializes itself to those values and
is enabled only when the rig has the capability so set CTCSS tone.